+++ app/src/main/java/com/example/charo_android/presentation/di/ViewModelModule.kt
@@ -3,7 +3,7 @@ package com.example.charo_android.presentation.di
 import com.example.charo_android.presentation.ui.home.viewmodel.HomeViewModel
 import com.example.charo_android.presentation.ui.main.SharedViewModel
 import com.example.charo_android.presentation.ui.more.viewmodel.MoreViewViewModel
-import com.example.charo_android.presentation.ui.mypage.viewmodel.FollowViewModel
+import com.example.charo_android.presentation.ui.follow.viewmodel.FollowViewModel
 import com.example.charo_android.presentation.ui.mypage.viewmodel.MyPageViewModel
 import com.example.charo_android.presentation.ui.search.viewmodel.SearchViewModel
 import com.example.charo_android.presentation.ui.setting.viewmodel.SettingViewModel
