+++ app/src/main/java/com/example/charo_android/presentation/ui/follow/FollowActivity.kt
@@ -1,12 +1,12 @@
-package com.example.charo_android.presentation.ui.mypage.follow
+package com.example.charo_android.presentation.ui.follow
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import com.example.charo_android.R
 import com.example.charo_android.databinding.ActivityFollowBinding
-import com.example.charo_android.presentation.ui.mypage.adapter.FollowViewPagerAdapter
-import com.example.charo_android.presentation.ui.mypage.viewmodel.FollowViewModel
+import com.example.charo_android.presentation.ui.follow.adapter.FollowViewPagerAdapter
+import com.example.charo_android.presentation.ui.follow.viewmodel.FollowViewModel
 import com.google.android.material.tabs.TabLayoutMediator
 import org.koin.androidx.viewmodel.ext.android.viewModel
