+++ app/src/main/res/layout/fragment_follower.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.mypage.follow.FollowerFragment">
+        tools:context=".presentation.ui.follow.FollowerFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv"
