+++ app/src/main/res/layout/fragment_following.xml
@@ -11,7 +11,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.mypage.follow.FollowingFragment">
+        tools:context=".presentation.ui.follow.FollowingFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv"
